% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{subtract}
\alias{subtract}
\title{Algebraic Subtraction}
\usage{
subtract(x, y)
}
\arguments{
\item{x}{A numeric vector from which \code{y} will be subtracted.}

\item{y}{A numeric vector to subtract.}
}
\value{
A numeric vector with the difference between \code{x} and \code{y}.
}
\description{
Finds the algebraic difference between two numeric vectors.
}
\examples{
subtract(c(10, 10), c(7, 20))
> [1]  3 -10
}
